Add asynchronous functions to the HuginBase::ImageCache class.
Use requestAsync(Small)Image to request an image later, using a background thread to load it. Use get(Small)ImageIfAvailable to get an EntryPtr to the image entry if it is in the cache, or 0 if not. This way the GUI doesn't need to flash place holders up if the image can be displayed quickly.

In Hugin, the ImageCache will wait for the background thread to load an image, then a wxEvent derivative delays signal handling until it is safe. This way the GUI can remain responsive while images are loaded without too much work.
Delay drawing of image thumbnails in the Images tab while the image loads.
Use placeholder textures in the Fast preview while the images load.