PTBatcherGUI: Simplify code for checking for changes
Because only one instance of PTBatcherGUI can ever write to batch queue file,
no need to fiddle with 2 changing file (.ptbt0 and .ptbt1). So now use only
one batch queue file (.ptbt). Therefor it is not necessary to constantly
checking the queue file for changes.
Instead of using a separate thread to constant start the checking for modified
project file now, use a simpler wxTimer.
This should reduce the load when running PTBatcher.